Distributed_and_Cloud_Computing-ICS432_01-Fall2021/Assignment-04/DanddankHomework4/DanddankNetflixApp/requirements.txt
Flask==1.1.2
firebase-admin
gunicorn==20.0.4
astroid == 2.3.3
pylint == 2.3.1
pytest >= 6.2.0
pytest-cov >= 2.4.0
pytest-localserver >= 0.4.1

cachecontrol >= 0.12.6
google-api-core[grpc] >= 1.22.1, < 3.0.0dev; platform.python_implementation != 'PyPy'
google-api-python-client >= 1.7.8
google-cloud-firestore >= 2.1.0; platform.python_implementation != 'PyPy'
google-cloud-storage >= 1.37.1